Refactor: use modules to define routes and tasks
Until now, routs were defined in the __init__.py file of the
project. Similarly, the celery tasks were scattered.

With this commit, FastAPI routs (i.e. endpoints) are defined in the
=router= directory, also, the celery tasks have a specific home.

Also, I have updated the README file to state that this project is not only
for python3.6.5 and offline use, it is a good modular example of using
Celery, FastAPI and SocketIO together.